LayoutTests/fast/parser/script-modify-page-outer.html

<!DOCTYPE html>
<html>
<script>

if (window.testRunner)
    testRunner.dumpAsText();

</script>
<body onload="document.body.appendChild(document.createTextNode('DONE'));">
<iframe src="resources/script-modify-page-inner.html"></iframe>
</body>
</html>
