LayoutTests/platform/ios-sim-deprecated/fast/parser/script-modify-page-outer-expected.txt

ALERT: should be outer: parser/script-modify-page-outer.html
ALERT: should be inner: parser/resources/script-modify-page-inner.html
DONE
